Added action plugin for generate_examples
