+++ roles/module_openapi_cloud/tasks/main.yaml
@@ -1,7 +1,8 @@
 ---
 - name: Generate schema for "{{ collection['name'] }}"
-  set_fact:
-    schema_generation: " {{ plugin['api_object_path'] | refresh_schema(plugin['resource']) }}"
+  ansible.content_builder.generate_cloud_schema:
+      api_object_path: "{{ plugin['api_object_path'] }}"
+      resource: "{{ plugin['resource'] }}"
   when:
     - ( plugin['action'] == 'generate_schema' ) or ( plugin['action'] == 'generate_all' )
     - plugin['name'] == 'amazon_cloud'
@@ -17,10 +18,8 @@
   when: ( plugin['action'] == 'generate_modules' ) or ( plugin['action'] == 'generate_all' )
 
 - name: Generate examples for "{{ collection['name'] }}"
-  ansible.builtin.command: python3 "{{ role_path }}"/templates/refresh_examples.py
-    --target-dir "{{ collection['path'] }}"
-  delegate_to: 127.0.0.1
-  run_once: true
+  ansible.content_builder.generate_cloud_examples:
+      target_dir: "{{ collection['path'] }}"
   changed_when: false
   when: ( plugin['action'] == 'generate_examples' ) or ( plugin['action'] == 'generate_all' )
